examples: Tier A + Tier B starter programs (39 in total)
Adds runnable example binaries under examples/ covering both the
fundamentals (Tier A: pinning, manifests, feeds, SOC/GSOC, ACT, PSS,
postage, offline helpers) and a set of small starter projects (Tier B:
swarm-paste, swarm-deploy, swarm-blog, swarm-fs, swarm-chat,
swarm-vault, swarm-share, swarm-pinner, swarm-feed-rss, swarm-relay,
swarm-keyring, swarm-cost-monitor) plus an examples/README.md that
groups everything by primitive and adds setup / caveats notes.

Also pulls in two small library fixes the example suite surfaced:

- src/debug/accounting.rs: RedistributionState.last_sample_duration_seconds
  is f64; Bee returns it as a fractional float (e.g. 37.96302), so the
  previous u64 typing failed deserialization. Eq is dropped from the
  derive since f64 doesn't implement it.
- examples/integration-check.rs: retry the PSS send a couple of times
  before giving up — Bee occasionally 500s when no peer is currently
  reachable in the target neighborhood.

Cargo.toml: scrypt + aes-gcm added as dev-dependencies for the
swarm-keyring example (passphrase-encrypted local keystore).

Verified with all 39 examples running end-to-end against a local Bee
2.7.2-rc1 node (no Sepolia traffic).

+++ src/debug/accounting.rs
@@ -86,7 +86,7 @@ pub struct PeerAccounting {
 
 /// Redistribution-state snapshot. Mirrors bee-go
 /// `RedistributionStateResponse`.
-#[derive(Clone, Debug, PartialEq, Eq, Default, Deserialize)]
+#[derive(Clone, Debug, PartialEq, Default, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct RedistributionState {
     /// Minimum gas funds to play a round.
@@ -110,8 +110,9 @@ pub struct RedistributionState {
     pub last_frozen_round: u64,
     /// Last round selected.
     pub last_selected_round: u64,
-    /// Last sample duration in seconds.
-    pub last_sample_duration_seconds: u64,
+    /// Last sample duration in seconds. Bee returns this as a
+    /// fractional float (e.g. `37.96302`), not an integer.
+    pub last_sample_duration_seconds: f64,
     /// Latest seen block.
     pub block: u64,
     /// Cumulative reward (PLUR).
